MLproject_Solar_Irradiance/dashweb/requirements.txt
dash
pandas
Pillow
plotly
tkinter